beckend-ml/requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
mysql-connector-python